% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.gsea-package.R
\docType{package}
\name{augere.gsea-package}
\alias{augere.gsea}
\alias{augere.gsea-package}
\title{augere.gsea: Automatic Generation of Gene Set Enrichment Analyses}
\description{
Implements pipelines for generating gene set enrichment analysis reports in the augere framework. This includes various competitive and self-contained gene sets from a variety of Bioconductor packages. Each pipeline function generates a self-contained Rmarkdown report with all of the steps required to reproduce the gene set enrichment analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/augere-bioinfo/augere.gsea}
  \item Report bugs at \url{https://github.com/augere-bioinfo/augere.gsea/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})

}
