% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGeneSetTestCommands.R
\name{generateGeneSetTestCommands}
\alias{generateGeneSetTestCommands}
\title{Commands for limma's mean-rank gene set test}
\usage{
generateGeneSetTestCommands(
  sets.name,
  stat.name,
  sign.name,
  alternative = c("mixed", "up", "down", "either"),
  seed = NULL,
  use.sqrt = FALSE,
  args = list()
)
}
\arguments{
\item{sets.name}{String containing the variable name for the list of gene sets.
Each set should be an integer vector of indices into the variable named by \code{is.sig.name}.}

\item{stat.name}{String containing the variable name of the numeric vector containing the test statistic for ranking genes.
The referenced vector should be of length equal to the number of genes.}

\item{sign.name}{String containing the variable name for the numeric vector of signed effect sizes (typically log-fold changes) for each gene.
The referenced vector should be of length equal to the number of genes.
It can be \code{NULL} if the statistics referenced by \code{stat.name} is already signed or \code{alternative="mixed"}.}

\item{alternative}{String specifying the alternative hypothesis. 
This should be one of:
\itemize{
\item \code{"mixed"}: test each set for enrichment of genes with the largest absolute values of the statistics. 
\item \code{"up"}: test each set for enrichment of genes with the largest statistics, typically up-regulated.
\item \code{"down"}: test each set for enrichment of genes with the smallest statistics, typically down-regulated.
\item \code{"either"}: test each set for enrichment of genes in either direction.
}}

\item{seed}{Seed for the random number generator.
If \code{NULL}, no seed setting is performed.}

\item{use.sqrt}{Boolean indicating whether to take the square root of the unsigned statistics when converting them to signed values.}

\item{args}{Named list of arguments to pass to \code{\link[limma]{geneSetTest}}.}
}
\value{
Character vector containing the commands required to run \code{geneSetTest}.
Upon evaluation, this produces a \link[S4Vectors]{DFrame} with one row per gene set and the following columns:
\itemize{
\item \code{NumGenes}, the number of genes in each set.
\item \code{PValue}, the p-value for enrichment of highly-ranked genes in this set.
\item \code{FDR}, the Benjamini-Hochberg-adjusted p-value.
\item \code{Direction}, whether the set is more enriched for up- or down-regulated genes.
Only reported for \code{alternative="either"}.
}
}
\description{
Generate R commands to run \code{\link[limma]{geneSetTest}} under various parametrizations.
}
\section{Direction of the alternative hypothesis}{

In general, it is expected that the statistics in \code{stat.name} are signed values from a differential expression analysis, e.g., t-statistics, z-scores.
Larger positive and negative values should correspond to stronger up- and down-regulation, respectively.
This ensures that \code{alternative="up"} or \code{alternative="down"} will have the expected behavior.

If the input statistics are unsigned, they are converted to their signed equivalents by multiplying it with the sign of the variable referenced by \code{sign}.
For F-statistics or chi-squared statistics, we can set \code{use.sqrt=TRUE} to convert them to t-statistics and z-scores, respectively.
This mimics the behavior of a test that produces signed test statistics.

For \code{alternative="either"}, each gene set is tested separately for enrichment of up- and down-regulated genes.
The reported \code{Direction} is defined based on the one with a lower p-value.
In contrast, \code{alternative="mixed"} does not consider the sign at all, and only cares about enrichment of significant genes of any direction in each set.
}

\examples{
cat(generateGeneSetTestCommands("sets", "stat", "sign"), sep="\n")
cat(generateGeneSetTestCommands("sets", "stat", "sign", alternative="either"), sep="\n")

}
\author{
Aaron Lun
}
