% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHypergeometricTestCommands.R
\name{generateHypergeometricTestCommands}
\alias{generateHypergeometricTestCommands}
\title{Commands for a hypergeometric test}
\usage{
generateHypergeometricTestCommands(
  sets.name,
  is.sig.name,
  sign.name,
  alternative = c("mixed", "up", "down", "either")
)
}
\arguments{
\item{sets.name}{String containing the variable name for the list of gene sets.
Each set should be an integer vector of indices into the variable named by \code{is.sig.name}.}

\item{is.sig.name}{String containing the variable name for the logical vector indicating which genes are significant.
The referenced vector should be of length equal to the number of genes.}

\item{sign.name}{String containing the variable name for the numeric vector of signed effect sizes (typically log-fold changes) for each gene.
The referenced vector should be of length equal to the number of genes.
It can be omitted for \code{alternative="mixed"}.}

\item{alternative}{String specifying the alternative hypothesis.
This should be one of:
\itemize{
\item \code{"mixed"} tests for enrichment of any significant genes in each set, regardless of their direction.
\item \code{"up"} tests for enrichment of up-regulated genes in each set.
\item \code{"down"} tests for enrichment of down-regulated genes in each set.
\item \code{"either"} tests for enrichment of either up- or down-regulated genes in each set.
}}
}
\value{
Character vector containing the commands required to perform a hypergeometric test.
Upon evaluation, this produces a \link[S4Vectors]{DFrame} with one row per gene set and the following columns:
\itemize{
\item \code{NumGenes}, the number of genes in each set. 
\item \code{NumSig}, the number of significant genes in each set, possibly after filtering for the desired sign.
\item \code{PValue}, the p-value for enrichment from the hypergeometric test.
\item \code{FDR}, the Benjamini-Hochberg-adjusted p-value.
\item \code{Direction}, whether the set is more enriched for up- or down-regulated genes.
Only reported for \code{alternative="either"}.
}
}
\description{
Generate R commands to run a hypergeometric test under various parameterizations.
}
\section{Direction of the alternative hypothesis}{

For \code{alternative="either"}, each gene set is tested separately for enrichment of up- and down-regulated genes.
The reported \code{Direction} is defined based on the one with a lower p-value.
In contrast, \code{alternative="mixed"} does not consider the sign at all, and only cares about enrichment of significant genes of any direction in each set.
}

\examples{
cat(generateHypergeometricTestCommands("sets", "is.sig", "sign"), sep="\n")
cat(generateHypergeometricTestCommands("sets", "is.sig", "sign", alternative="either"), sep="\n")

}
\author{
Aaron Lun
}
