% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runContrast.R
\name{runContrast}
\alias{runContrast}
\title{Differential gene set analyses with contrasts}
\usage{
runContrast(
  x,
  sets,
  groups,
  comparison,
  covariates = NULL,
  block = NULL,
  subset.factor = NULL,
  subset.levels = NULL,
  subset.groups = TRUE,
  design = NULL,
  contrast = NULL,
  dc.block = NULL,
  robust = TRUE,
  quality = TRUE,
  trend = FALSE,
  methods = c("fry", "camera"),
  assay = 1,
  annotation = NULL,
  metadata = NULL,
  output.dir = "contrast",
  author = NULL,
  dry.run = FALSE,
  save.results = TRUE
)
}
\arguments{
\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} object containing a count matrix where genes and samples are in rows and columns, respectively.
Alternatively, the output of \code{\link[augere.core]{wrapInput}} that refers to a SummarizedExperiment.}

\item{sets}{A list or \link[IRanges]{CharacterList} of character vectors.
Each vector represents a gene set and contains the identifiers for genes in that set.
Identifiers should be consistent with the row names of \code{x}.
The list may also be named with the names of the sets.}

\item{groups}{String specifying the \code{\link[SummarizedExperiment]{colData}(x)} column containing the grouping factor of interest,
see \code{\link[augere.de]{processSimpleDesignMatrix}} for more details.
This may be \code{NULL} for experimental designs with no groups, e.g., \code{covariates} only.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{comparison}{Character vector of length no greater than 2.
For length 2, this specifies the groups to be compared, whereas for length 1, this specifies the covariate to test. 
See \code{\link[augere.de]{processSimpleComparisons}} for more details.

Alternatively, a named character vector with no more than 2 unique names,
see \code{\link[augere.de]{processSimpleComparisons}} for more details.

Unlike \code{\link[augere.de]{runVoom}}, a list of vectors is not accepted.}

\item{covariates}{Character vector specifying the \code{\link[SummarizedExperiment]{colData}(x)} columns containing continuous covariates of interest,
see \code{\link[augere.de]{processSimpleDesignMatrix}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{block}{Character vector specifying the \code{\link[SummarizedExperiment]{colData}(x)} columns containing additional (uninteresting) blocking factors,
see \code{\link[augere.de]{processSimpleDesignMatrix}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{subset.factor}{String specifying the \code{\link[SummarizedExperiment]{colData}(x)} column containing the factor to use for subsetting.}

\item{subset.levels}{Vector containing the levels of the \code{subset.factor} to be retained.}

\item{subset.groups}{Boolean indicating whether to automatically subset the dataset to only those samples assigned to groups in \code{comparisons}.
Setting this to \code{TRUE} sacrifices some residual degrees of freedom for greater robustness against variability in irrelevant groups.
Ignored if \code{design} and \code{contrasts} are provided.
Also ignored if \code{covariates} is provided, as all samples are informative for a continuous covariate.}

\item{design}{Matrix, function, or formula specifying the experimental design, see \code{\link[augere.de]{processCustomDesignMatrix}} for details.
If this and \code{contrasts} are specified, \code{groups}, \code{block}, \code{covariates}, \code{comparisons} and \code{subset.groups} are ignored.}

\item{contrast}{String, function or vector specifying a custom contrast,
see \code{?\link[augere.de]{processCustomContrasts}} for more details.

Unlike \code{\link[augere.de]{runVoom}}, a list of contrasts is not accepted.}

\item{dc.block}{String specifying the blocking factor to use in \code{\link[limma]{duplicateCorrelation}}.
Typically used for uninteresting factors that cannot be used in \code{block} as they are confounded with the factors of interest.
No additional blocking is performed if \code{NULL}.}

\item{robust}{Boolean indicating whether robust empirical Bayes shrinkage should be used in \code{\link[limma]{eBayes}}.
Setting this to \code{TRUE} sacrifices some precision for improved robustness against genes with extreme variances.}

\item{quality}{Boolean indicating whether quality weighting should be performed.
This reduces the influence of low-quality samples at the cost of more computational work.}

\item{trend}{Boolean indicating whether variances should be shrunk towards a trend in \code{\link[limma]{eBayes}}.
Usually unnecessary as the observation weights already account for the mean-variance relationship.}

\item{methods}{Character vector specifying the methods to run.
This should contain at least one of the following:
\itemize{
\item \code{"mroast"}, which \code{\link[limma]{mroast}} from the \pkg{limma} package.
This is a self-contained gene set test where the null hypothesis is that the genes in the set are not differentially expressed.
\item \code{"fry"}, which calls \code{\link[limma]{fry}} from the \pkg{limma} package.
This is also a self-contained gene set test that is a fast approximation of \code{mroast}.
\item \code{"camera"}, which calls \code{\link[limma]{camera}} from the \pkg{limma} package.
This is a competitive gene set test where the null hypothesis is that the genes in the set are not more DE than the genes outside the set.
\item \code{"romer"}, which calls \code{\link[limma]{romer}} from the \pkg{limma} package.
This is also a competitive test that takes a different approach to accounting for correlations between genes in the same set.
}}

\item{assay}{String or integer specifying the assay of \code{x} containing the count matrix.}

\item{annotation}{Character vector specifying the columns of \code{\link[S4Vectors]{mcols}(sets)} to store in each result DataFrames.}

\item{metadata}{Named list of additional metadata to store alongside each result.}

\item{output.dir}{String containing the path to an output directory in which to write the Rmarkdown file and save results.}

\item{author}{Character vector containg the names of the authors.
If \code{NULL}, defaults to the current user.}

\item{dry.run}{Boolean indicating whether to perform a dry run.
This generates the Rmarkdown report in \code{output.dir} but does not execute the analysis.}

\item{save.results}{Boolean indicating whether the results should be saved to file.}
}
\value{
A Rmarkdown report named \code{report.Rmd} is written inside \code{output.dir}. 
This contains all commmands used to reproduce the analysis. 

If \code{dry.run=FALSE}, a list of \link[S4Vectors]{DataFrame}s is returned where each DataFrame contains the enrichment results for a \code{method}.
Each row corresponds to a gene set in \code{sets} and is named accordingly.
All DataFrames are guaranteed to have (at least) the following fields:
\itemize{
\item \code{NumGenes}, the number of genes in the set (after removing genes that were not tested).
\item \code{Direction}, the net direction of the change in expression within the set (either \code{"up"} or \code{"down"}).
\item \code{PValue}, the p-value for enrichment in each gene set.
\item \code{FDR}, the Benjamini-Hochberg-adjusted p-value. 
}
Additional fields may be present for specific methods.

If \code{save.results=TRUE}, the results are saved in a \code{results} subdirectory of \code{output.dir}.

If \code{dry.run=FALSE}, only the report is created, and \code{NULL} is returned.
}
\description{
Analyze a RNA-seq dataset for differential expression across a collection of gene sets.
This requires access to the original values, unlike \code{\link{runPrecomputed}}.
}
\details{
Some of the methods involve randomization, so for full reproducibility, users should call \code{\link{set.seed}} before running \code{runContrast}.

Note that, even if the user does not call \code{\link{set.seed}},
\code{runContrast} will automatically insert \code{\link{set.seed}} statements into the Rmarkdown report 
prior to any GSEA functions that involve randomization.
Each \code{set.seed} call has a hard-coded seed to ensure that future compilation of the generated report will give the same result.
However, different calls to \code{runContrast} will use different (randomly selected) seeds to avoid systematic biases.
Thus, if full reproducibility of \code{runContrast} is required, users should set the seed themselves before calling \code{runContrast}.
}
\examples{
x <- augere.de::loadExampleDataset()
all.genes <- rownames(x)

sets <- list(
    A = sample(all.genes, 92),
    B = sample(all.genes, 212),
    C = sample(all.genes, 12),
    D = sample(all.genes, 38),
    E = sample(all.genes, 55)
)

output.dir <- tempfile()
res <- runContrast(
    x,
    sets,
    group="dex", 
    comparison=c("trt", "untrt"),
    output=output.dir
)

res
list.files(output.dir, recursive=TRUE)

}
\author{
Aaron Lun
}
