% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augere.gsea-package.R
\name{augere.gsea-package}
\alias{augere.gsea-package}
\title{augere.gsea}
\description{
Implement gene set enrichment analyses in the augere framework.
Pipeline functions execute gene set enrichment analyses with packages like limma, fgsea or goseq,
also generating a self-contained Rmarkdown report to reproduce each analysis.
}
