% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runContrast.R
\name{runContrast}
\alias{runContrast}
\title{Differential gene set analyses with contrasts}
\usage{
runContrast(
  se,
  collection,
  groups,
  comparison,
  block = NULL,
  covariates = NULL,
  assay = 1,
  dup.cor = NULL,
  robust = TRUE,
  qualities = TRUE,
  trend = FALSE,
  subset.factor = NULL,
  subset.levels = NULL,
  subset.groups = TRUE,
  design = NULL,
  contrast = NULL,
  anno = NULL,
  methods = c("fry", "camera"),
  output = "gsea-contrast",
  metadata = NULL,
  author = NULL,
  dry.run = FALSE,
  save.all = TRUE,
  rand.seed = NULL
)
}
\arguments{
\item{groups}{String specifying the \code{\link[SummarizedExperiment]{colData}(x)} column containing the grouping factor of interest,
see \code{\link[augere.de]{processSimpleDesignMatrix}} for more details.
This may be \code{NULL} for experimental designs with no groups, e.g., \code{covariates} only.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{comparison}{Character vector of length no greater than 2.
For length 2, this specifies the groups to be compared, whereas for length 1, this specifies the covariate to test. 
See \code{\link[augere.de]{processSimpleComparisons}} for more details.

Alternatively, a named character vector with no more than 2 unique names,
see \code{\link[augere.de]{processSimpleComparisons}} for more details.

Unlike \code{\link[augere.de]{runVoom}}, a list of vectors is not accepted.}

\item{block}{Character vector specifying the \code{\link[SummarizedExperiment]{colData}(x)} columns containing additional (uninteresting) blocking factors,
see \code{\link[augere.de]{processSimpleDesignMatrix}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{covariates}{Character vector specifying the \code{\link[SummarizedExperiment]{colData}(x)} columns containing continuous covariates of interest,
see \code{\link[augere.de]{processSimpleDesignMatrix}} for more details.
Ignored if \code{design} and \code{contrasts} are provided.}

\item{assay}{String or integer specifying the assay of \code{x} containing the count matrix.}

\item{robust}{Boolean indicating whether robust empirical Bayes shrinkage should be used in \code{\link[limma]{eBayes}}.
Setting this to \code{TRUE} sacrifices some precision for improved robustness against genes with extreme variances.}

\item{trend}{Boolean indicating whether variances should be shrunk towards a trend in \code{\link[limma]{eBayes}}.
Usually unnecessary as the observation weights already account for the mean-variance relationship.}

\item{subset.factor}{String specifying the \code{\link[SummarizedExperiment]{colData}(x)} column containing the factor to use for subsetting.}

\item{subset.levels}{Vector containing the levels of the \code{subset.factor} to be retained.}

\item{subset.groups}{Boolean indicating whether to automatically subset the dataset to only those samples assigned to groups in \code{comparisons}.
Setting this to \code{TRUE} sacrifices some residual degrees of freedom for greater robustness against variability in irrelevant groups.
Ignored if \code{design} and \code{contrasts} are provided.
Also ignored if \code{covariates} is provided, as all samples are informative for a continuous covariate.}

\item{design}{Matrix, function, or formula specifying the experimental design, see \code{\link[augere.de]{processCustomDesignMatrix}} for details.
If this and \code{contrasts} are specified, \code{groups}, \code{block}, \code{covariates}, \code{comparisons} and \code{subset.groups} are ignored.}

\item{contrast}{String, function or vector specifying a custom contrast,
see \code{?\link[augere.de]{processCustomContrasts}} for more details.

Unlike \code{\link[augere.de]{runVoom}}, a list of contrasts is not accepted.}

\item{methods}{Character vector specifying the methods to run.
This should contain at least one of the following:
\itemize{
\item \code{"mroast"}, which \code{\link[limma]{mroast}} from the \pkg{limma} package.
This is a self-contained gene set test where the null hypothesis is that the genes in the set are not differentially expressed.
\tiem \code{"fry"}, which calls \code{\link[limma]{fry}} from the \pkg{limma} package.
This is also a self-contained gene set test that is a fast approximation of \code{mroast}.
\item \code{"camera"}, which calls \code{\link[limma]{camera}} from the \pkg{limma} package.
This is a competitive gene set test where the null hypothesis is that the genes in the set are not more DE than the genes outside the set.
\item \code{"romer"}, which calls \code{\link[limma]{romer}} from the \pkg{limma} package.
This is also a competitive test that takes a different approach to accounting for correlations between genes in the same set.
}}

\item{metadata}{Named list of additional metadata to store alongside each result.}

\item{author}{Character vector containg the names of the authors.
If \code{NULL}, defaults to the current user.}

\item{dry.run}{Boolean indicating whether to perform a dry run.
This generates the Rmarkdown report in \code{output.dir} but does not execute the analysis.}
}
\value{
A Rmarkdown report named \code{report.Rmd} is written inside \code{output.dir}. 
This contains all commmands used to reproduce the analysis. 

If \code{dry.run=FALSE}, a list of \link[S4Vectors]{DataFrame}s is returned where each DataFrame contains the enrichment results for a \code{method}.
Each row corresponds to a gene set in \code{sets} and is named accordingly.
All DataFrames are guaranteed to have (at least) the following fields:
\itemize{
\item \code{NumGenes}, the number of genes in the set (after removing genes that were not tested).
\item \code{Direction}, the net direction of the change in expression within the set (either \code{"up"} or \code{"down"}).
\item \code{PValue}, the p-value for enrichment in each gene set.
\item \code{FDR}, the Benjamini-Hochberg-adjusted p-value. 
}
Additional fields may be present for specific methods.

If \code{save.results=TRUE}, the results are saved in a \code{results} subdirectory of \code{output.dir}.

If \code{dry.run=FALSE}, only the report is created, and \code{NULL} is returned.
}
\description{
Analyze a RNA-seq dataset for differential expression across a collection of gene sets.
This requires access to the original values, unlike \code{\link{runGseaPrecomputed}}.
}
\examples{
all.keys <- as.character(seq_len(1000))

library(SummarizedExperiment)
mu <- 2^runif(length(all.keys), 3, 10)
se <- SummarizedExperiment(list(counts=
    matrix(rnbinom(length(mu) * 10, mu=mu, size=10), nrow=length(mu))),
    colData=DataFrame(g=paste0("group", gl(2, 5)))
)
rownames(se) <- all.keys

# Mocking up some gene sets.
collection <- IRanges::CharacterList(
    A=sample(all.keys, 100),
    B=sample(all.keys, 500),
    C=sample(all.keys, 200),
    D=sample(all.keys, 20),
    E=sample(all.keys, 10)
)

# Running the gene set tests.
proj <- tempfile()
out <- runGSEAContrast(se,
    collection,
    group="g", 
    comparison=c("group2", "group1"),
    output=proj)

list.files(proj, recursive=TRUE)
out

}
\author{
Aaron Lun
}
